// src/app/Models/GetFlowViewShiftRoteGetApiDataClass.ts
// Generated by https://quicktype.io

export interface GetFlowViewShiftRoteGetApiDataClass {
    EmpID1:                string;
    EmpName1:              string;
    DeptName1:             string;
    EmpID2:                string;
    EmpName2:              string;
    DeptName2:             string;
    FlowViewShiftRoteDate: FlowViewShiftRoteDate[];
    ShiftRoteName:         string;
    ShiftRoteCount:        number;
    Key:                   string;
    ProcessFlowID:         number;
    FormName:              string;
    AppEmpID:              string;
    AppEmpName:            string;
    AppDeptName:           string;
    ManageEmpName:         string;
    RealManageEmpName:     string;
    State:                 string;
    Take:                  boolean;
    AbsFlowApps:           any;
    AbscFlowApps:          any;
    CardFlowApps:          any;
    ShiftRoteFlowApps:     any;
    Handle:            boolean;
    TransSign:             boolean;
}


export interface FlowViewShiftRoteDate {
    ShiftRoteDate: string;
    RoteID1:       number;
    RoteCode1:     string;
    RoteName1:     string;
    RoteID2:       number;
    RoteCode2:     string;
    RoteName2:     string;
}